code cleaning.

+++ src/main/java/com/example/LearnHowToUserGit/services/ConfigService.java
@@ -5,13 +5,13 @@ import org.springframework.stereotype.Service;
 @Service
 public class ConfigService {
 
-    int timeToLive = 86400;
+    long timeToLive = 86400;
     String authTokenSecret = "UTF-8";
 
     public String getAuthTokenSecret() {
         return authTokenSecret;
     }
-    public int getTimeToLive() {
+    public long getTimeToLive() {
         return timeToLive;
     }
 }
